fixing profiles listing

+++ modules/zsh/install.sh
@@ -11,23 +11,25 @@ mkdir -p "$CUSTOM_DIR/themes"
 # install oh-my-zsh
 if [ ! -d "$ZSH_DIR" ]; then
     echo "Installing Oh My Zsh..."
-    git clone https://github.com/ohmyzsh/ohmyzsh.git "$ZSH_DIR"
+    git clone --depth 1 https://github.com/ohmyzsh/ohmyzsh.git "$ZSH_DIR"
 fi
 
 # install powerlevel10k
 if [ ! -d "$CUSTOM_DIR/themes/powerlevel10k" ]; then
     echo "Installing Powerlevel10k..."
-    git clone https://github.com/romkatv/powerlevel10k.git \
+    git clone --depth 1 https://github.com/romkatv/powerlevel10k.git \
         "$CUSTOM_DIR/themes/powerlevel10k"
 fi
 
 install_plugin () {
-    name="$1"
-    repo="$2"
+    local name="$1"
+    local repo="$2"
 
-    if [ ! -d "$CUSTOM_DIR/plugins/$name" ]; then
+    if [ -d "$CUSTOM_DIR/plugins/$name/.git" ]; then
+        git -C "$CUSTOM_DIR/plugins/$name" pull --quiet
+    else
         echo "Installing plugin: $name"
-        git clone "$repo" "$CUSTOM_DIR/plugins/$name"
+        git clone --depth 1 "$repo" "$CUSTOM_DIR/plugins/$name"
     fi
 }
 
@@ -40,25 +42,26 @@ https://github.com/marlonrichert/zsh-autocomplete
 install_plugin zsh-autosuggestions \
 https://github.com/zsh-users/zsh-autosuggestions
 
-# catppuccin theme
-TMP_DIR=$(mktemp -d)
-git clone https://github.com/catppuccin/zsh-fsh.git "$TMP_DIR"
-cp "$TMP_DIR/themes/catppuccin-mocha.ini" \
-"$CUSTOM_DIR/plugins/fast-syntax-highlighting/themes"
-rm -rf "$TMP_DIR"
+# install catppuccin theme
+THEME_FILE="$CUSTOM_DIR/plugins/fast-syntax-highlighting/themes/catppuccin-mocha.ini"
 
-ensure_zshrc() {
+if [ ! -f "$THEME_FILE" ]; then
+    TMP_DIR=$(mktemp -d)
 
-if [ ! -e "$HOME/.zshrc" ] && [ ! -L "$HOME/.zshrc" ]; then
-    echo "[zsh] Creating placeholder .zshrc"
+    git clone --depth 1 https://github.com/catppuccin/zsh-fsh.git "$TMP_DIR"
 
-    cat > "$HOME/.zshrc" <<'EOF'
-# placeholder created by setup-env
-EOF
+    cp "$TMP_DIR/themes/catppuccin-mocha.ini" \
+    "$CUSTOM_DIR/plugins/fast-syntax-highlighting/themes"
+
+    rm -rf "$TMP_DIR"
 fi
-}
-ensure_zshrc
+
 # change default shell
-if command -v zsh >/dev/null && [ "$SHELL" != "$(command -v zsh)" ]; then
-    chsh -s "$(command -v zsh)"
+if command -v zsh >/dev/null; then
+    ZSH_PATH="$(command -v zsh)"
+
+    if [ "$SHELL" != "$ZSH_PATH" ]; then
+        echo "Setting default shell to zsh..."
+        chsh -s "$ZSH_PATH" || echo "Could not change shell automatically"
+    fi
 fi
